@@ -1,22 +1,25 @@
 {
   "name": "urn-lib",
-  "version": "1.2.0",
-  "description": "Creating, parsing, and validating ",
-  "main": "dist/urn-lib.umd.js",
-  "jsnext:main": "dist/urn-lib.es2015.js",
-  "module": "dist/urn-lib.es2015.js",
+  "version": "2.0.0-alpha.0",
+  "description": "Creating, parsing, and validating",
+  "main": "dist/urn-lib.cjs.js",
+  "browser": "dist/urn-lib.umd.js",
+  "types": "dist/main.d.ts",
   "author": {
     "name": "Cory Mawhorter",
     "email": "cory.mawhorter@gmail.com",
     "url": "http://www.mawhorter.net/"
   },
   "scripts": {
-    "prebuild": "npm run lint && npm run test",
-    "build": "rollup -c",
-    "postbuild": "node ./scripts/assert-build.js",
-    "lint": "eslint ./src",
-    "coverage": "nyc --reporter=html npm run test",
-    "test": "mocha -r ./test/lib/helper test"
+    "build": "webpack -c webpack.config.js",
+    "prepublishOnly": "npm test && npm run build",
+    "pretest": "npm run lint",
+    "test": "npm run testonly",
+    "testonly": "mocha ./test",
+    "lintsource": "eslint --fix ./src",
+    "linttests": "eslint --fix ./test",
+    "lint": "npm run lintsource && npm run linttests",
+    "upgrade": "npm-check --update"
   },
   "repository": {
     "type": "git",
@@ -27,25 +30,25 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "babel-plugin-external-helpers": "^6.22.0",
-    "babel-preset-es2015": "^6.22.0",
-    "babel-register": "^6.22.0",
-    "builtin-modules": "^1.1.1",
-    "chai": "^4.2.0",
-    "eslint": "^6.8.0",
-    "mocha": "^7.0.1",
-    "nyc": "^15.0.0",
-    "reify": "^0.20.12",
-    "rollup": "^0.41.4",
-    "rollup-plugin-babel": "^2.7.1",
-    "rollup-plugin-commonjs": "^7.0.0",
-    "rollup-plugin-node-resolve": "^2.0.0",
-    "source-map-support": "^0.4.15"
+    "@types/chai": "^4.2.14",
+    "@types/mocha": "^8.0.4",
+    "@types/node": "^14.14.10",
+    "@typescript-eslint/eslint-plugin": "^4.8.2",
+    "@typescript-eslint/parser": "^4.8.2",
+    "chai": "^4.3.4",
+    "eslint": "^7.14.0",
+    "eslint-plugin-import": "^2.22.1",
+    "mocha": "^8.2.1",
+    "npm-check": "^5.9.2",
+    "source-map-loader": "^2.0.1",
+    "source-map-support": "^0.5.19",
+    "ts-loader": "^8.0.11",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.1.2",
+    "webpack": "^5.31.1",
+    "webpack-cli": "^4.2.0"
   },
-  "babel": {
-    "ignore": false,
-    "presets": [
-      "es2015"
-    ]
-  }
+  "files": [
+    "dist/"
+  ]
 }
