@@ -1,11 +1,9 @@
-'use strict';
-
-var URN = require('../src/main.js');
+import { expect } from 'chai';
+import { RFC2141 } from '../src/main';
 
 describe('RFC2141', function() {
   it('should export an instance for parsing RFC2141 urns', function() {
-    var rfc2141 = URN.RFC2141;
-    var canned = [
+    const canned = [
       'urn:isbn:0451450523',
       'urn:isan:0000-0000-9E59-0000-O-0000-0000-2',
       'urn:ISSN:0167-6423',
@@ -16,10 +14,10 @@ describe('RFC2141', function() {
       'urn:nbn:de:bvb:19-146642',
       'urn:lex:eu:council:directive:2010-03-09;2010-19-UE',
     ];
-    canned.forEach(function(urn) {
-      var parsed = rfc2141.parse(urn);
-      var formatted = rfc2141.format(parsed);
-      expect(formatted).to.equal(urn);
-    });
+    for (const urn of canned) {
+      const parsed = RFC2141.parse(urn);
+      const formatted = RFC2141.format(parsed as any);
+      expect(formatted).to.equal(urn, `failed: ${urn}`);
+    }
   });
 });
